+++ src/db/user/user.queries.js
@@ -5,6 +5,4 @@ export const SQL_QUERIES = {
   FIND_CHARACTER_BY_USER_ID_AND_CLASS: 'SELECT * FROM Characters WHERE user_id = ? AND job_id = ?',
   INSERT_CHARACTER:
     'INSERT INTO Characters (user_id, name, job_id, job_name, hp, mp, attack, defense) VALUES (?, ?, ?, ?, ?, ?, ?, ?)',
-
-  UPDATE_COORDINATE: 'UPDATE coordinates_backup SET x = ?, y = ? WHERE device_id = ?',
 };
